Fixes issue with duplicate variants
When selecting variation groups and variant products as part of the export, the current state retrieves all variants of the master products instead of just the variants associated with the variation group creating duplicate variant products in the export feed. This change splits out the check for master and variation groups to retrieve only the relevant children.